tools/requirements.txt
Pillow
libconf